Move getShadow* to ShadowBaseNode
